+++ integrations/dense/test_dpr.py
@@ -32,7 +32,7 @@ class TestDpr(unittest.TestCase):
         self.threads = 16
         self.batch_size = self.threads * 32
 
-        half_cores = multiprocessing.cpu_count() / 2
+        half_cores = int(multiprocessing.cpu_count() / 2)
         # If server supports more threads, then use more threads.
         # As a heuristic, use up half up available CPU cores.
         if half_cores > self.threads:
